*has added new logic: replace, separate;

+++ DVHtmlToModels_Example/DVHtmlToModels_Example/Models/PriceObject.h
@@ -10,6 +10,6 @@
 
 @interface PriceObject : NSObject
 @property (nonatomic) int discount;
-@property (nonatomic) int price;
+@property (nonatomic) double price;
 @property (nonatomic, strong) NSString *currency;
 @end
